Get update sınıfları düzeltildi frontend tasarlandı email servisi değiştirildi

+++ Application/Features/Command/AdvertisementTransactionHandlers/UpdateAdvertisement/UpdateAdvertisementCommandRequest.cs
@@ -12,12 +12,14 @@ namespace Application.Features.Command.AdvertisementTransactionHandlers.UpdateAd
     {
         [JsonIgnore]
         public string? OwnerId { get; set; }
+
         [JsonIgnore]
         public string? AdvertisementId { get; set; }
+
         public string Title { get; set; }
         public string AdvertisementAddress { get; set; }
-        public double? Latitude { get; set; }
-        public double? Longitude { get; set; }
+        public decimal? Latitude { get; set; }
+        public decimal? Longitude { get; set; }
         public string ImgUrl { get; set; }
     }
 }
